+++ linkhighlighter.js
@@ -2,10 +2,10 @@
  *  linkhighlighter.js - Tiny library to highlight active link on web page.
  *
  *  @author     Михаил Драгункин <contact@unsektor.com>
- *  @url        https://github.com/unsektor/linkhighlighter.js/
+ *  @url        https://github.com/unsektor/linkhighlighter/
  *  @license    ISC
  *  @since      June 25, 2017
- *  @ver 	    1.0.0
+ *  @ver        1.1.0
  */
 
 ;(function (root, factory) {
@@ -27,7 +27,7 @@
          */
         var highlightScope = function (scopeElement) {
             var scopeHighlightClass = scopeElement.getAttribute('data-lh-class') || 'g-lh-active';
-            var anchorElementList = scopeElement.querySelectorAll('[data-lh]');
+            var anchorElementList = scopeElement.querySelectorAll('a[data-lh]');
 
             anchorElementList.forEach(function (anchorElement) {
                 // this code is not complicated too strict to validate option value is correct.
@@ -36,25 +36,24 @@
 
                 if ((anchorElementOption === 'match-domain' && anchorElement.host === window.location.host) ||
                     (anchorElementOption === 'match-uri' && anchorElement.href === window.location.href) ||
-                    (anchorElementOption === 'match-partial' && (new RegExp('^' + anchorElement.href, 'i')).test(window.location.href))
+                    (anchorElementOption === 'match-partial' && window.location.href.toLowerCase().startsWith(anchorElement.href.toLowerCase()))
                 ) {
                     anchorElement.classList.add(scopeHighlightClass);
-                } else {
-                    // this code removes class name added past time.
-                    anchorElement.classList.remove(scopeHighlightClass);
+                    return;
                 }
+
+                // remove class name added earlier
+                anchorElement.classList.remove(scopeHighlightClass);
             });
         };
 
-        /**
-         * Main function
-         */
         this.highlight = function () {
             var scopeElementList = document.querySelectorAll('[data-lh-scope]');
 
             // if at least one scope was not defined, root element as only scope will be used.
             if (0 === scopeElementList.length) {
-                scopeElementList = document.querySelectorAll('html');
+                highlightScope(document.querySelector('html'));
+                return;
             }
 
             scopeElementList.forEach(highlightScope);