+++ CHANGELOG.md
@@ -8,11 +8,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 
-+ Tests
-+ DOM & URI changes watch
-+ Default class name, URI match option configuration 
-+ Browsers compatibility table
+- [ ] Cover with unit-tests
+- [ ] DOM & URI changes watch
+- [ ] Default class name, URI match option configuration
 
-## 1.0.0 - 2017-06-25
+## [1.1.0] - 2025-11-22
+### Changed
+
+- [x] refactoring: switch `match-partial` check method from regexp to substring
+
+### Fixed
+
+- [x] fix: clarify `data-lh` attribute element query selector to anchor element
+- [x] fix: update documentation issues
+
+## [1.0.0] - 2017-06-25
 
 - Initial release
+
+[1.1.0]: https://github.com/unsektor/linkhighlighter/compare/1.0.0...1.1.0
+[1.0.0]: https://github.com/unsektor/linkhighlighter/releases/tag/1.0.0